# Conflicts:
#	gulpfile.js
#	package-lock.json
#	package.json
#	src/web/assets/admintable/dist/js/app.js
#	src/web/assets/admintable/dist/js/app.js.map
#	src/web/assets/admintable/dist/js/chunk-vendors.js.map
#	src/web/assets/admintable/package-lock.json
#	src/web/assets/admintable/package.json
#	src/web/assets/craftsupport/dist/CraftSupportWidget.css.map
#	src/web/assets/d3/D3Asset.php
#	src/web/assets/dashboard/dist/dashboard.css.map
#	src/web/assets/datepickeri18n/DatepickerI18nAsset.php
#	src/web/assets/deprecationerrors/dist/deprecator.css.map
#	src/web/assets/edittransform/dist/transforms.css.map
#	src/web/assets/edituser/dist/account.css.map
#	src/web/assets/edituser/dist/profile.css.map
#	src/web/assets/elementresizedetector/ElementResizeDetectorAsset.php
#	src/web/assets/fabric/FabricAsset.php
#	src/web/assets/fileupload/FileUploadAsset.php
#	src/web/assets/garnish/GarnishAsset.php
#	src/web/assets/generalsettings/dist/rebrand.css.map
#	src/web/assets/graphiql/dist/graphiql.js
#	src/web/assets/installer/dist/install.css.map
#	src/web/assets/jquerypayment/JqueryPaymentAsset.php
#	src/web/assets/login/dist/login.css.map
#	src/web/assets/matrixsettings/dist/MatrixConfigurator.min.js
#	src/web/assets/matrixsettings/dist/MatrixConfigurator.min.js.map
#	src/web/assets/picturefill/PicturefillAsset.php
#	src/web/assets/plugins/dist/plugins.css.map
#	src/web/assets/pluginstore/PluginStoreAsset.php
#	src/web/assets/pluginstore/dist/css/app.css
#	src/web/assets/pluginstore/dist/js/app.js
#	src/web/assets/pluginstore/dist/js/app.js.map
#	src/web/assets/pluginstore/dist/js/chunk-vendors.js
#	src/web/assets/pluginstore/dist/js/chunk-vendors.js.map
#	src/web/assets/pluginstoreoauth/dist/pluginstore-oauth-callback.css.map
#	src/web/assets/prismjs/PrismJsAsset.php
#	src/web/assets/queuemanager/dist/queue-manager.min.js.map
#	src/web/assets/qunit/QunitAsset.php
#	src/web/assets/routes/dist/routes.css
#	src/web/assets/routes/dist/routes.css.map
#	src/web/assets/selectize/SelectizeAsset.php
#	src/web/assets/systemmessages/dist/system_messages.css.map
#	src/web/assets/timepicker/TimepickerAsset.php
#	src/web/assets/updater/dist/Updater.min.js
#	src/web/assets/updater/dist/Updater.min.js.map
#	src/web/assets/updater/dist/update.css.map
#	src/web/assets/updates/dist/updates.css.map
#	src/web/assets/userpermissions/dist/user-permissions.css.map
#	src/web/assets/utilities/dist/utilities.css.map
#	src/web/assets/velocity/VelocityAsset.php
#	src/web/assets/xregexp/XregexpAsset.php

+++ src/console/controllers/utils/PruneRevisionsController.php
@@ -27,15 +27,22 @@ class PruneRevisionsController extends Controller
     /**
      * @var int The maximum number of revisions an element can have.
      */
-    public $maxRevisions;
+    public int $maxRevisions;
+
+    /**
+     * @var bool Whether this is a dry run.
+     * @since 3.7.9
+     */
+    public $dryRun = false;
 
     /**
      * @inheritdoc
      */
-    public function options($actionID)
+    public function options($actionID): array
     {
         $options = parent::options($actionID);
         $options[] = 'maxRevisions';
+        $options[] = 'dryRun';
         return $options;
     }
 
@@ -46,7 +53,7 @@ class PruneRevisionsController extends Controller
      */
     public function actionIndex(): int
     {
-        if ($this->maxRevisions === null) {
+        if (!isset($this->maxRevisions)) {
             $this->maxRevisions = $this->prompt('What is the max number of revisions an element can have?', [
                 'default' => Craft::$app->getConfig()->getGeneral()->maxRevisions,
                 'validator' => function($input) {
@@ -59,18 +66,18 @@ class PruneRevisionsController extends Controller
         $this->stdout('Finding elements with too many revisions ... ');
         $elements = (new Query())
             ->select([
-                'id' => 's.sourceId',
+                'id' => 's.canonicalId',
                 's.count',
                 'type' => (new Query())
                     ->select(['type'])
                     ->from([Table::ELEMENTS])
-                    ->where(new Expression('[[id]] = [[s.sourceId]]')),
+                    ->where(new Expression('[[id]] = [[s.canonicalId]]')),
             ])
             ->from([
                 's' => (new Query())
-                    ->select(['sourceId', 'count' => 'COUNT(*)'])
+                    ->select(['canonicalId', 'count' => 'COUNT(*)'])
                     ->from(['r' => Table::REVISIONS])
-                    ->groupBy(['sourceId'])
+                    ->groupBy(['canonicalId'])
                     ->having(['>', 'COUNT(*)', $this->maxRevisions]),
             ])
             ->all();
@@ -93,19 +100,31 @@ class PruneRevisionsController extends Controller
             /** @var ElementInterface|string $elementType */
             $elementType = $element['type'];
             $deleteCount = $element['count'] - $this->maxRevisions;
+
             $this->stdout('- ' . $elementType::displayName() . " {$element['id']} ({$deleteCount} revisions) ... ");
+
             $extraRevisions = $elementType::find()
                 ->revisionOf($element['id'])
                 ->siteId('*')
                 ->unique()
-                ->anyStatus()
+                ->status(null)
                 ->orderBy(['num' => SORT_DESC])
                 ->offset($this->maxRevisions)
                 ->all();
-            foreach ($extraRevisions as $extraRevision) {
-                $elementsService->deleteElement($extraRevision, true);
+
+            if (!$this->dryRun) {
+                foreach ($extraRevisions as $extraRevision) {
+                    $elementsService->deleteElement($extraRevision, true);
+                }
             }
-            $this->stdout('done' . PHP_EOL, Console::FG_GREEN);
+
+            $this->stdout('done', Console::FG_GREEN);
+
+            if (count($extraRevisions) !== $deleteCount) {
+                $this->stdout(' (found ' . count($extraRevisions) . ')', Console::FG_RED);
+            }
+
+            $this->stdout(PHP_EOL);
         }
 
         $this->stdout(PHP_EOL . 'Finished pruning revisions' . PHP_EOL . PHP_EOL, Console::FG_GREEN);
